refactor: break up accounts into separate go file
